pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-infrastructure-git-tagging"
version = "0.0.1a6"
description = "Infrastructure layer for pythoneda-artifact/git-tagging"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-infrastructure/git-tagging"
packages = [{include = "pythonedaartifactinfrastructuregittagging"}]

[tool.poetry.dependencies]
dbus-next = "^0.2.3"
grpcio = "^1.41.0"
python = "^3.8"
pythoneda-base = "^0.0.1a15"
pythoneda-artifact-event-git-tagging = "^0.0.1a1"
pythoneda-artifact-git-tagging = "^0.0.1a5"
pythoneda-infrastructure-base = "^0.0.1a11"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
